Refactor Chernova 2017 scripts to not duplicate all the functions
