+++ scripts/Chernova 2017 - MgO.py
@@ -16,76 +16,9 @@ import matplotlib.pyplot as plt
 
 import matplotlib
 matplotlib.use("TkAgg")
-#
-# def tauc_lorentz_analytic(eV, Eg, A, E0, C):
-#
-#     #
-#     # Epsilon 2
-#     #
-#
-#     eps2 = np.zeros(eV.shape)
-#     for i, e in enumerate(eV):
-#         if e > Eg:
-#             eps2[i] = (1 / e) * ( (A * E0 * C * (e - Eg)**2)
-#                              / ((e**2 - E0**2)**2 + C**2 * e**2))
-#
-#     #
-#     # Epsilon 1 closed form
-#     #
-#     eps1 = np.zeros(eV.shape)
-#     for i, e in enumerate(eV):
-#
-#         alpha_ln = (Eg**2 - E0**2)*e*2 + Eg**2*C**2 - E0**2 * (E0**2 + 3*Eg**2)
-#         alpha_atan = (e**2 - E0**2) * (E0**2 +Eg**2) + Eg**2 * C**2
-#
-#         alpha = np.sqrt(4 * E0**2 - C**2)
-#         gamma = np.sqrt(E0**2 - C**2/2)
-#
-#         xi4 = (e**2 - gamma**2)**2 + alpha**2 * C**2 / 4
-#
-#         eps1[i] = (
-#             + (1/2)*(A/np.pi)*(C/xi4)*(alpha_ln/(alpha*E0))*np.log2(
-#                 (E0**2 + Eg**2 + alpha*Eg) / (E0**2 + Eg**2 - alpha*Eg)
-#             )
-#             - (A / (np.pi * xi4)) * (alpha_atan/E0) * (np.pi
-#                                                        - np.arctan((2*Eg+alpha)/C)
-#                                                        + np.arctan((-2*Eg+alpha)/C))
-#             + 2 * (A*E0*C)/(np.pi*xi4) * (Eg*(e**2-gamma**2)*(np.pi + 2*np.arctan((gamma**2 - Eg**2)/(alpha*C))))
-#             - 2 * ((A*E0*C)/(np.pi * xi4))*((e**2 + Eg**2)/e)*np.log2(np.abs(e-Eg)/(e+Eg))
-#             + 2 * (A*E0*C)/(np.pi*xi4)*Eg*np.log2((np.abs(e-Eg)*(e+Eg))/np.sqrt((E0**2 - Eg**2)**2+Eg**2*C**2))
-#         )
-#
-#     return eps1, eps2
-#
-
-
-def TaucLorentz_KK(eV, E0, A, C, Eg):
-    eps_2 = np.zeros(eV.shape)
-
-    for j, e in enumerate(eV):
-        if e > Eg:
-            eps_2[j] = (1. / e) * (
-                    (A * E0 * C * (e - Eg) ** 2) /
-                    (
-                        (e ** 2 - E0 ** 2) ** 2 + C ** 2 * e ** 2
-                    )
-            )
-
-    eps_1 = kk_integral_maclaurin(eV, eps_2)
-
-    return eps_1, eps_2
-
-
-def Lorentz(eV, A, FWHM, Eg):
-    Br = FWHM #/ (2 * np.sqrt(np.log(2)))
-    # epsilon = [A*e**2 / (Eg**2-e**2-1j*e*Br) for e in eV]
-    # return np.real(epsilon), np.imag(epsilon)
-
-    eps_2 = [A * Br**2 * Eg * e / ((Eg**2 - e**2)**2 + Br**2*e**2) for e in eV]
-    eps_1 = [A * Br * Eg * (Eg**2 - e**2) / ((Eg**2 - e**2)**2 + Br**2*e**2) for e in eV]
-    return np.asarray(eps_1), np.asarray(eps_2)
 
 def generate_epsilon():
+    auxfuncs = __import__("Chernova 2017 - Aux Funcs")
 
     UV_E = 13.7
     UV_Amplitude = 47
@@ -113,7 +46,7 @@ def generate_epsilon():
     plt.figure(0)
     # Tauc-Lorentz
     for i in range(len(E_TL)):
-        eps_1_TL, eps_2_TL = TaucLorentz_KK(eV, E_TL[i], A_TL[i], C_TL[i], Eg_TL[i])
+        eps_1_TL, eps_2_TL = auxfuncs.taucLorentz_KK(eV, E_TL[i], A_TL[i], C_TL[i], Eg_TL[i])
         eps1 += eps_1_TL
         eps2 += eps_2_TL
 
@@ -122,7 +55,7 @@ def generate_epsilon():
     # Lorentz oscillators
     #
     for i in range(len(Lorentz_Eg)):
-        eps_1_lor, eps_2_lor = Lorentz(eV, Lorentz_Amplitude[i], Lorentz_FWHM[i], Lorentz_Eg[i])
+        eps_1_lor, eps_2_lor = auxfuncs.lorentz(eV, Lorentz_Amplitude[i], Lorentz_FWHM[i], Lorentz_Eg[i])
         eps1 += eps_1_lor
         eps2 += eps_2_lor
 
@@ -140,31 +73,6 @@ def generate_epsilon():
     plt.show()
 
     return eV, epsilon
-
-
-def kk_integral_maclaurin(f, eps2):
-    #
-    # KK integral
-    #
-    df = f[1] - f[0]
-    eps1 = np.zeros(f.shape)
-    for i, e in enumerate(f):
-        prefactor = (2 / np.pi) * 2 * df
-
-        maclaurin_sum = 0
-        if i % 2 == 0:
-            js = [z for z in range(f.shape[-1])[1::2]]
-        else:
-            js = [z for z in range(f.shape[-1])[0::2]]
-
-        for j in js:
-            maclaurin_sum += (1. / 2.) * (
-                    eps2[j] / (f[j] - e) +
-                    eps2[j] / (f[j] + e)
-            )
-        eps1[i] = prefactor * maclaurin_sum
-
-    return eps1
 
 
 if __name__ == "__main__":
